[LOGGER] put common defines into runtime_util folder, ran tests to confirm code still functioning
